harmonydb-backend/requirements.txt
django
djangorestframework
psycopg2-binary
requests  # Optional for AI
Pillow    # Optional for images
djangorestframework-simplejwt
django-cors-headers
python-dotenv